Add settings for enhanced completion menu
Only insert the longest match of all strings rather than the
whole first match. Also, open the menu for one single match as well.

Confirm a selection with <Enter> rather than <Ctrl-y>.

Simulate the down arrow to always have the first item selected and
be able to narrow the matches as you type.

Signed-off-by: hjan <hoeppner.jan@xurv.org>